+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - (Enhancement) Add support for Resource Profile (`fmc_resource_profiles`) resource and data source
 - (Enhancement) Add support for multi-instance (`fmc_chassis_*`) resources and data sources. Tune `fmc_device_subinterface` and `fmc_device_etherchannel_interface` to support multi-instance logical devices.
+- (Enhancement) Add support for Bridge Group Interface (BVI)  (`fmc_device_bridge_group_interface`)
 
 ## 2.0.0-beta5